+++ crud/truck.py
@@ -4,13 +4,14 @@ from fastapi import HTTPException
 from schemas import truck as truck_schema
 
 
-def read(db: Session, id: int):
-    query = db.query(models.Truck).filter(models.Truck.id == id)
+def read(db: Session, id: int, user_id: int):
+    query = db.query(models.Truck).filter_by(id=id, user_id=user_id)
     return query.first()
 
 
-def create(db: Session, truck: truck_schema.TruckCreate):
-    truck_db = models.Truck(**truck.dict())
+def create(db: Session, truck: truck_schema.TruckCreate, user_id: int):
+    # TODO: фильтрация для tuck_type других юзеров
+    truck_db = models.Truck(**truck.dict(), user_id=user_id)
     db.add(truck_db)
     try:
         db.commit()
@@ -18,13 +19,13 @@ def create(db: Session, truck: truck_schema.TruckCreate):
         raise HTTPException(status_code=400, detail=str(e))
 
 
-def read_all(db: Session):
-    query = db.query(models.Truck)
+def read_all(db: Session, user_id: int):
+    query = db.query(models.Truck).filter_by(user_id=user_id)
     return query.all()
 
 
-def delete(db: Session, truck_id: int):
-    db.query(models.Truck).filter_by(id=truck_id).delete()
+def delete(db: Session, truck_id: int, user_id: int):
+    db.query(models.Truck).filter_by(id=truck_id, user_id=user_id).delete()
     try:
         db.commit()
     except Exception as e:
@@ -32,8 +33,8 @@ def delete(db: Session, truck_id: int):
         raise HTTPException(status_code=400, detail=str(e))
 
 
-def update(db: Session, truck_id: int, truck: truck_schema.TruckBase):
-    db.query(models.Truck).filter_by(id=truck_id).update({**truck.dict()})
+def update(db: Session, truck_id: int, truck: truck_schema.TruckBase, user_id: int):
+    db.query(models.Truck).filter_by(id=truck_id, user_id=user_id).update({**truck.dict()})
     try:
         db.commit()
     except Exception as e:
